kernel: sched: fix redundancy in idle loop
We cast `cpu` to a long, even though it gets casted to a (void *)
right after. It is useless to process this cast.

+++ kernel/sched/idle.c
@@ -209,7 +209,7 @@ DEFINE_PER_CPU(bool, cpu_dead_idle);
  */
 static void cpu_idle_loop(void)
 {
-	int cpu = smp_processor_id();
+	unsigned int cpu = smp_processor_id();
 
 	while (1) {
 		/*
@@ -231,7 +231,7 @@ static void cpu_idle_loop(void)
 
 			if (cpu_is_offline(cpu)) {
 				rcu_cpu_notify(NULL, CPU_DYING_IDLE,
-					       (void *)(long)cpu);
+					       (void *)cpu);
 				smp_mb(); /* all activity before dead. */
 				this_cpu_write(cpu_dead_idle, true);
 				arch_cpu_idle_dead();
